+++ .coveragerc
@@ -2,7 +2,12 @@
 [run]
 branch = True
 source = gldas
-omit = */_version.py
+# omit = bad_file.py
+
+[paths]
+source =
+    src/
+    */site-packages/
 
 [report]
 # Regexes for lines to exclude from consideration
