--- .gitattributes
@@ -1,1 +1,0 @@
-gldas/_version.py export-subst
