+++ CHANGELOG.rst
@@ -5,6 +5,14 @@ Changelog
 Unreleased
 ==========
 
+-
+
+Version 0.6
+===========
+
+- New package structure
+- Drop python 2 support
+- Allow subsetting by bbox
 - Add gldas land mask to package (no download necessary)
 
 Version 0.5
