--- doc_requirements.txt
@@ -1,6 +1,0 @@
-mock
-numpy
-pygeobase
-datedown
-trollsift
-sphinx==1.6.6
