+++ docs/changelog.rst
@@ -1,0 +1,2 @@
+.. _changes:
+.. include:: ../CHANGELOG.rst
