--- docs/changes.rst
@@ -1,2 +1,0 @@
-.. _changes:
-.. include:: ../CHANGES.rst
