--- docs/environment.yml
@@ -1,22 +1,0 @@
-#this file is not used at the moment
-name: gldas-docs
-channels:
-  - conda-forge
-dependencies:
-- matplotlib
-- netcdf4
-- pandas
-- basemap
-- pygrib
-- scipy
-- pyresample
-- sphinx
-- python=3.6
-- pygrib
-- pip:
-    - pygeobase
-    - https://software.ecmwf.int/wiki/download/attachments/56664858/ecmwf-api-client-python.tgz#egg=ecmwf_api_client
-    - pynetcf
-    - repurpose
-    - pytesmo
-    - pygeogrids
