--- gldas/__init__.py
@@ -1,6 +1,0 @@
-import pkg_resources
-
-try:
-    __version__ = pkg_resources.get_distribution(__name__).version
-except:
-    __version__ = 'unknown'
