--- readthedocs.yml
@@ -1,4 +1,0 @@
-python:
-   version: 3
-   setup_py_install: true
-requirements_file: doc_requirements.txt
