--- requirements.txt
@@ -1,12 +1,0 @@
-pyproj
-pygeogrids
-pygrib
-numpy
-pygeobase
-datedown>=0.3
-trollsift
-netcdf4
-pyresample
-repurpose
-pynetcf
-more_itertools
