+++ setup.cfg
@@ -1,44 +1,83 @@
+# This file is used to configure your project.
+# Read more about the various options under:
+# http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
+
 [metadata]
 name = gldas
-summary = Readers and converters for data from the GLDAS Noah Land Surface Model.
-author = Package authors
-author-email = christoph.paulik@geo.tuwien.ac.at
-license = new-bsd
-home-page = http://...
-description-file = README.rst
+description = Readers and converters for data from the GLDAS Noah Land Surface Model.
+author = TU Wien
+author-email = remote.sensing@geo.tuwien.ac.at
+license = mit
+long-description = file: README.rst
+long-description-content-type = text/x-rst; charset=UTF-8
+url = https://github.com/TUW-GEO/gldas
+project-urls =
+    Documentation = https://gldas.readthedocs.io/en/latest/
+# Change if running only on Windows, Mac or Linux (comma-separated)
+platforms = any
 # Add here all kinds of additional classifiers as defined under
 # https://pypi.python.org/pypi?%3Aaction=list_classifiers
-classifiers = Development Status :: 4 - Beta
-              Programming Language :: Python
+classifiers =
+    Development Status :: 4 - Beta
+    Programming Language :: Python
 
-[entry_points]
+[options]
+zip_safe = False
+packages = find:
+include_package_data = True
+package_dir =
+    =src
+# DON'T CHANGE THE FOLLOWING LINE! IT WILL BE UPDATED BY PYSCAFFOLD!
+setup_requires = pyscaffold>=3.2a0,<3.3a0
+# Add here dependencies of your project (semicolon/line-separated), e.g.
+install_requires =  pyproj
+                    pygeogrids
+                    pygrib
+                    numpy
+                    pygeobase
+                    datedown>=0.3
+                    trollsift
+                    netcdf4
+                    pyresample
+                    repurpose
+                    pynetcf
+
+# The usage of test_requires is discouraged, see `Dependency Management` docs
+# tests_require = pytest-cov; coverage; pytest;
+# Require a specific Python version, e.g. Python 2.7 or >= 3.4
+# python_requires = >=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*
+
+[options.packages.find]
+where = src
+exclude =
+    tests
+
+[options.extras_require]
+# Add here additional requirements for extra features, to install with:
+# `pip install gldas[PDF]` like:
+# PDF = ReportLab; RXP
+# Add here test requirements (semicolon/line-separated)
+testing =
+    pytest-cov<=2.9
+    coverage==4.5.2
+    pytest==3.8.2
+
+[options.entry_points]
 # Add here console scripts like:
+# console_scripts =
+#     script_name = gldas.module:function
+# For example:
 console_scripts =
     gldas_download = gldas.download:run
     gldas_repurpose = gldas.reshuffle:run
-# For example:
-# console_scripts =
-#     fibonacci = gldas.skeleton:run
-# as well as other entry_points.
-
-
-[files]
-# Add here 'data_files', 'packages' or 'namespace_packages'.
-# Additional data files are defined as key value pairs of source and target:
-packages =
-    gldas
-# data_files =
-#    share/gldas_docs = docs/*
-
-[extras]
-# Add here additional requirements for extra features, like:
-# PDF =
-#    ReportLab>=1.2
-#    RXP
+# And any other entry points, for example:
+# pyscaffold.cli =
+#     awesome = pyscaffoldext.awesome.extension:AwesomeExtension
 
 [test]
 # py.test options when running `python setup.py test`
-addopts = tests
+# addopts = --verbose
+extras = True
 
 [tool:pytest]
 # Options for py.test:
@@ -48,9 +87,14 @@ addopts = tests
 addopts =
     --cov gldas --cov-report term-missing
     --verbose
+norecursedirs =
+    dist
+    build
+    .tox
+testpaths = tests
 
 [aliases]
-docs = build_sphinx
+dists = bdist_wheel
 
 [bdist_wheel]
 # Use this option if your package is pure-python
@@ -58,20 +102,27 @@ universal = 1
 
 [build_sphinx]
 source_dir = docs
-build_dir = docs/_build
-
-[pbr]
-# Let pbr run sphinx-apidoc
-autodoc_tree_index_modules = True
-# autodoc_tree_excludes = ...
-# Let pbr itself generate the apidoc
-# autodoc_index_modules = True
-# autodoc_exclude_modules = ...
-# Convert warnings to errors
-# warnerrors = True
+build_dir = build/sphinx
 
 [devpi:upload]
-# Options for the devpi: PyPI serer and packaging tool
+# Options for the devpi: PyPI server and packaging tool
 # VCS export must be deactivated since we are using setuptools-scm
 no-vcs = 1
 formats = bdist_wheel
+
+[flake8]
+# Some sane defaults for the code style checker flake8
+exclude =
+    .tox
+    build
+    dist
+    .eggs
+    docs/conf.py
+
+[pyscaffold]
+# PyScaffold's parameters when the project was created.
+# This will be used when updating. Do not change!
+version = 3.2.3
+package = gldas
+extensions =
+    no_skeleton
