+++ setup.py
@@ -1,23 +1,23 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
     Setup file for gldas.
+    Use setup.cfg to configure your project.
 
-    This file was generated with PyScaffold 2.5rc4, a tool that easily
-    puts up a scaffold for your new Python project. Learn more under:
-    http://pyscaffold.readthedocs.org/
+    This file was generated with PyScaffold 3.2.3.
+    PyScaffold helps you to put up the scaffold of your new Python project.
+    Learn more under: https://pyscaffold.org/
 """
-
 import sys
+
+from pkg_resources import VersionConflict, require
 from setuptools import setup
 
-
-def setup_package():
-    needs_sphinx = {'build_sphinx', 'upload_docs'}.intersection(sys.argv)
-    sphinx = ['sphinx'] if needs_sphinx else []
-    setup(setup_requires=['six', 'pyscaffold>=2.5a0,<2.6a0'] + sphinx,
-          use_pyscaffold=True)
+try:
+    require('setuptools>=38.3')
+except VersionConflict:
+    print("Error: version of setuptools is too old (<38.3)!")
+    sys.exit(1)
 
 
 if __name__ == "__main__":
-    setup_package()
+    setup(use_pyscaffold=True)
