--- test-requirements.txt
@@ -1,6 +1,0 @@
-# Add requirements only needed for your unittests and during development here.
-# They will be installed automatically when running `python setup.py test`.
-# ATTENTION: Don't remove pytest-cov and pytest as they are needed.
-pytest-cov
-coverage==4.5.2
-pytest==3.8.2
