+++ tests/conftest.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+    Dummy conftest.py for gldas.
+
+    If you don't know what this is for, just leave it empty.
+    Read more about conftest.py under:
+    https://pytest.org/latest/plugins.html
+"""
+
+# import pytest
